fix(generator): fix _#fromFields validation, double-super bug, reorganize bindings
Generator fixes:
- _#fromFields now validates all required constructor params are matched
  by fields before emitting, eliminating 20 missing_required_argument errors
- Zero-fields early path also checks for required params (ElevatedButton etc.)
- mustCallSuper non-void pattern stores superResult to avoid calling super twice
- plugin_emitter accepts configurable bindingsImportPrefix (no longer hardcoded)

Structure:
- Bindings reorganized into per-package subdirectories:
  bindings/ui/, bindings/painting/, bindings/widgets/, bindings/material/, etc.
- Test imports updated for new paths
- Fix Uint32List→Uint64List in widget_bridge_test (64-bit ISA migration)

+++ packages/dartic_flutter/test/widget_bridge_test.dart
@@ -7,9 +7,9 @@ import 'package:dartic/src/bytecode/module.dart';
 import 'package:dartic/src/runtime/class_info.dart';
 import 'package:dartic/src/runtime/object.dart';
 import 'package:dartic_flutter/dartic_flutter.dart';
-import 'package:dartic_flutter/src/bindings/stateless_widget_bindings.g.dart';
-import 'package:dartic_flutter/src/bindings/stateful_widget_bindings.g.dart';
-import 'package:dartic_flutter/src/bindings/state_bindings.g.dart';
+import 'package:dartic_flutter/src/bindings/widgets/stateless_widget_bindings.g.dart';
+import 'package:dartic_flutter/src/bindings/widgets/stateful_widget_bindings.g.dart';
+import 'package:dartic_flutter/src/bindings/widgets/state_bindings.g.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
@@ -51,7 +51,7 @@ DarticDispatch _dispatchWith(Map<String, Object? Function(List<Object?>)> handle
     final nameIdx = nameIndices[entry.key]!;
     final proto = DarticFuncProto(
       funcId: funcId,
-      bytecode: Uint32List(0),
+      bytecode: Uint64List(0),
       valueRegCount: 0,
       refRegCount: 0,
       paramCount: 0,
